property: more on php8.0

+++ property/inc/class.uiproject.inc.php
@@ -2016,7 +2016,7 @@ JS;
 			if (isset($values['reserve']) && $values['reserve'] != 0)
 			{
 				$reserve_remainder	 = $values['reserve'] - $values['deviation'];
-				$remainder_percent	 = number_format(($reserve_remainder / $values['reserve']) * 100, 2, $this->decimal_separator, '');
+				$remainder_percent	 = number_format((float)($reserve_remainder / $values['reserve']) * 100, 2, $this->decimal_separator, '');
 				$values['sum']		 = $values['sum'] + $values['reserve'];
 			}
 
@@ -2093,13 +2093,13 @@ JS;
 			if (isset($values['reserve']) && $values['reserve'] != 0)
 			{
 				$reserve_remainder	 = $values['reserve'] - $values['deviation'];
-				$remainder_percent	 = number_format(($reserve_remainder / $values['reserve']) * 100, 2, $this->decimal_separator, '.');
+				$remainder_percent	 = number_format((float)($reserve_remainder / $values['reserve']) * 100, 2, $this->decimal_separator, '.');
 				$values['sum']		 = $values['sum'] + $values['reserve'];
 			}
 
 			$value_remainder = $values['sum'] - $sum_actual_cost - $sum_oblications;
-			$values['sum']	 = number_format($values['sum'], 0, $this->decimal_separator, '.');
-			$value_remainder = number_format($value_remainder, 0, $this->decimal_separator, '.');
+			$values['sum']	 = number_format((float)$values['sum'], 0, $this->decimal_separator, '.');
+			$value_remainder = number_format((float)$value_remainder, 0, $this->decimal_separator, '.');
 
 			$s_amount_in	 = 0;
 			$s_amount_out	 = 0;
@@ -2121,12 +2121,12 @@ JS;
 					array('key' => 'entry_date', 'label' => lang('entry date'), 'sortable' => false),
 					array('key'			 => 'amount_in', 'label'			 => lang('amount in'), 'sortable'		 => false,
 						'className'		 => 'right', 'formatter'		 => 'JqueryPortico.FormatterAmount0',
-						'value_footer'	 => number_format($s_amount_in, 0, $this->decimal_separator, '.')),
+						'value_footer'	 => number_format((float)$s_amount_in, 0, $this->decimal_separator, '.')),
 					array('key'		 => 'from_project', 'label'		 => lang('from project'), 'sortable'	 => true,
 						'className'	 => 'right', 'formatter'	 => 'project_link'),
 					array('key'			 => 'amount_out', 'label'			 => lang('amount out'), 'sortable'		 => false,
 						'className'		 => 'right', 'formatter'		 => 'JqueryPortico.FormatterAmount0',
-						'value_footer'	 => number_format($s_amount_out, 0, $this->decimal_separator, '.')),
+						'value_footer'	 => number_format((float)$s_amount_out, 0, $this->decimal_separator, '.')),
 					array('key'		 => 'to_project', 'label'		 => lang('to project'), 'sortable'	 => true,
 						'formatter'	 => 'project_link'),
 					array('key' => 'remark', 'label' => lang('remark'), 'sortable' => true)
@@ -2140,19 +2140,19 @@ JS;
 					array('key' => 'month', 'label' => lang('month'), 'sortable' => false),
 					array('key'			 => 'budget', 'label'			 => lang('budget'), 'sortable'		 => false,
 						'className'		 => 'right',
-						'formatter'		 => 'JqueryPortico.FormatterAmount0', 'value_footer'	 => number_format($s_budget, 0, $this->decimal_separator, '.')),
+						'formatter'		 => 'JqueryPortico.FormatterAmount0', 'value_footer'	 => number_format((float)$s_budget, 0, $this->decimal_separator, '.')),
 					array('key'			 => 'sum_oblications', 'label'			 => lang('sum orders'), 'sortable'		 => false,
 						'className'		 => 'right', 'formatter'		 => 'JqueryPortico.FormatterAmount0',
-						'value_footer'	 => number_format($sum_oblications, 0, $this->decimal_separator, '.')),
+						'value_footer'	 => number_format((float)$sum_oblications, 0, $this->decimal_separator, '.')),
 					array('key'			 => 'actual_cost', 'label'			 => lang('actual cost'), 'sortable'		 => false,
 						'className'		 => 'right', 'formatter'		 => 'JqueryPortico.FormatterAmount0',
-						'value_footer'	 => number_format($s_actual_cost, 0, $this->decimal_separator, '.')),
+						'value_footer'	 => number_format((float)$s_actual_cost, 0, $this->decimal_separator, '.')),
 					array('key'			 => 'diff', 'label'			 => lang('difference'), 'sortable'		 => false,
 						'className'		 => 'right',
-						'formatter'		 => 'JqueryPortico.FormatterAmount0', 'value_footer'	 => number_format($s_diff, 0, $this->decimal_separator, '.')),
+						'formatter'		 => 'JqueryPortico.FormatterAmount0', 'value_footer'	 => number_format((float)$s_diff, 0, $this->decimal_separator, '.')),
 					array('key'			 => 'deviation_period', 'label'			 => lang('deviation'), 'sortable'		 => false,
 						'className'		 => 'right', 'formatter'		 => 'JqueryPortico.FormatterAmount0',
-						'value_footer'	 => number_format($s_deviation, 0, $this->decimal_separator, '.')),
+						'value_footer'	 => number_format((float)$s_deviation, 0, $this->decimal_separator, '.')),
 					array('key'		 => 'deviation_acc', 'label'		 => lang('deviation') . '::' . lang('accumulated'),
 						'sortable'	 => false, 'className'	 => 'right', 'formatter'	 => 'JqueryPortico.FormatterAmount0'),
 					array('key'		 => 'deviation_percent_period', 'label'		 => lang('deviation') . '::' . lang('percent'),
